chore: move ProgressBar component in PlayerUI folder
